Reduce some cyclic imports by migrating some common types and exceptions to _types.py and _exceptions.py, use absolute imports
