+++ xclim/core/__init__.py
@@ -2,4 +2,6 @@
 
 from __future__ import annotations
 
-from . import missing
+from xclim.core import missing
+from xclim.core._exceptions import *
+from xclim.core._types import *
