+++ xclim/core/dataflags.py
@@ -15,19 +15,14 @@ from inspect import signature
 import numpy as np
 import xarray
 
-from ..indices.generic import binary_ops
-from ..indices.run_length import suspicious_run
-from .calendar import climatological_mean_doy, within_bnds_doy
-from .formatting import update_xclim_history
-from .units import convert_units_to, declare_units, infer_context, str2pint
-from .utils import (
-    VARIABLES,
-    InputKind,
-    MissingVariableError,
-    Quantified,
-    infer_kind_from_parameter,
-    raise_warn_or_log,
-)
+from xclim.core._exceptions import MissingVariableError, raise_warn_or_log
+from xclim.core._types import VARIABLES, Quantified
+from xclim.core.calendar import climatological_mean_doy, within_bnds_doy
+from xclim.core.formatting import update_xclim_history
+from xclim.core.units import convert_units_to, declare_units, infer_context, str2pint
+from xclim.core.utils import InputKind, infer_kind_from_parameter
+from xclim.indices.generic import binary_ops
+from xclim.indices.run_length import suspicious_run
 
 _REGISTRY = {}
