+++ xclim/core/options.py
@@ -12,8 +12,8 @@ from typing import Callable
 
 from boltons.funcutils import wraps
 
-from .locales import _valid_locales
-from .utils import ValidationError, raise_warn_or_log
+from xclim.core._exceptions import ValidationError, raise_warn_or_log
+from xclim.core.locales import _valid_locales
 
 METADATA_LOCALES = "metadata_locales"
 DATA_VALIDATION = "data_validation"
