+++ xclim/indices/__init__.py
@@ -2,15 +2,15 @@
 
 from __future__ import annotations
 
-from ._agro import *
-from ._anuclim import *
-from ._conversion import *
-from ._hydrology import *
-from ._multivariate import *
-from ._simple import *
-from ._synoptic import *
-from ._threshold import *
-from .fire import (
+from xclim.indices._agro import *
+from xclim.indices._anuclim import *
+from xclim.indices._conversion import *
+from xclim.indices._hydrology import *
+from xclim.indices._multivariate import *
+from xclim.indices._simple import *
+from xclim.indices._synoptic import *
+from xclim.indices._threshold import *
+from xclim.indices.fire import (
     cffwis_indices,
     drought_code,
     duff_moisture_code,
