+++ xclim/testing/helpers.py
@@ -15,8 +15,8 @@ from dask.callbacks import Callback
 
 import xclim
 import xclim.testing.utils as xtu
-from xclim.core import calendar
-from xclim.core.utils import VARIABLES
+from xclim.core import VARIABLES
+from xclim.core.calendar import percentile_doy
 from xclim.indices import (
     longwave_upwelling_radiation_from_net_downwelling,
     shortwave_upwelling_radiation_from_net_downwelling,
@@ -47,10 +47,10 @@ def generate_atmos(
     ) as ds:
         rsus = shortwave_upwelling_radiation_from_net_downwelling(ds.rss, ds.rsds)
         rlus = longwave_upwelling_radiation_from_net_downwelling(ds.rls, ds.rlds)
-        tn10 = calendar.percentile_doy(ds.tasmin, per=10)
-        t10 = calendar.percentile_doy(ds.tas, per=10)
-        t90 = calendar.percentile_doy(ds.tas, per=90)
-        tx90 = calendar.percentile_doy(ds.tasmax, per=90)
+        tn10 = percentile_doy(ds.tasmin, per=10)
+        t10 = percentile_doy(ds.tas, per=10)
+        t90 = percentile_doy(ds.tas, per=90)
+        tx90 = percentile_doy(ds.tasmax, per=90)
 
         ds = ds.assign(
             rsus=rsus,
